data-generator/requirements.txt
kafka-python-ng==2.2.3
faker==20.1.0